Add product option cost
We have added a new label and a bubble to the option weight field.  This is to add clarity for admins to input the weight change.  All shipping extensions and the checkout page are provisioned to add the change to the base weight.

On the new option Cost field  we have determined admins prefer to use the total option cost rather than the differential  as it is quicker and easier input from their invoicing,  rather than having to calculate the difference as is done in the weight field.  We have add a bubble to that field to explain the input.