Request add cost column method
Hi core devs,

Sorry for the changes. I change the add cost column from the upgrade.php file instead of using the sql due to we provide a patch for current v1.3.0 users in our site.

I am concern that when v1.3.0 users that already load the patch upgrade to v1.3.1 will get an error during the cart upgrade.