Add store id to customer class
Hi core devs,

In Multi-store setup we find that customer can login to different store that customer is not assigned to the other store.  So I have added the store_id to the customer class in order they are only able to login to the store that they registered on.