Separate guest and logged in user order cancellation URL (#1630)
* Separate guest and logged in user order cancellation URL

I found out that in the account/invoice order cancellation URL is using order_id instead of ot for guest, so i copied the code from fast checkout.

* add shipping_method as selector in fast checkout payment and shipping selector

We were adding some radio button in the shipping method section via hook and found out that the page was refreshed when we click the radio button. So i add the shipping_method as the selectorin fast checkout payment and shipping selector.