Update act on behalf to work for multi-store
Hi core devs,
i found out that the act on behalf for multi-store is not working correctly. i am asked to login but it continues to fail. it is the same on subfolder and subdomain for multi-store

it seems the issue is caused by the incorrect store_id

i propose the fix, i have tested it.