Update act on behalf to work for multi-store (#1744)
* Add additional hooks to order_details.tpl

Add additional hooks to order_details.tpl

* Update act on behalf to work for multi-store

Hi core devs,
i found out that the act on behalf for multi-store is not working correctly. i am asked to login but it continues to fail. it is the same on subfolder and subdomain for multi-store

it seems the issue is caused by the incorrect store_id

i propose the fix, i have tested it.

* Revert "Update act on behalf to work for multi-store"

This reverts commit 1920332272bf24e53b511f4a387e62a3a4577afc.

* update multi-store act on behalf

Hi core devs,
i found out that the act on behalf for multi-store is not working correctly in v1.4.2.1. i am asked to login but it continues to fail. it is the same on subfolder and subdomain for multi-store

it seems the issue is caused by the incorrect store_id

i propose the fix, i have tested it